Caroline (#1)
* organized and cleaned up room descriptions

also added character definitions and lightly edited some of the room descriptions. Room names needed to be changed so that Inform could understand character names

* locker locking

added verb definitions for unlocking a locker and hiding the contents of the locker type container

* implimented closing locker

upon leaving a room, a locker will close.

* added clues

right now they are floating offstage. need to add rules for when to bring them onstage.

* set up weapons

only one is implimented. easy to do more.